+++ config/routes.rb
@@ -10,4 +10,9 @@ Searls::Auth::Engine.routes.draw do
   post "login/verify", to: "verifications#create"
   get "login/verify_token", to: "verifications#create", as: :verify_token
   patch "login/resend_verification", to: "verifications#resend", as: :resend_verification
+
+  get "password/reset", to: "requests_password_resets#show", as: :password_reset_request
+  post "password/reset", to: "requests_password_resets#create"
+  get "password/reset/edit", to: "resets_passwords#show", as: :password_reset_edit
+  patch "password/reset", to: "resets_passwords#update", as: :password_reset_update
 end